Initial release of GUI version
This is the initial version of the GUI enables SAR simulator. The app is called from SARSimulator_v4. Other functions such as addTank and addMig can be downloaded in the main branch of this repository and are needed to run the simulation. bpsk_mod and costas_code must be downloaded from this branch because they are updated versions of the original.