Final Version Upload w/GUI
This is the final version of the code that includes a GUI. There are still a few errors in the code, but everything works most of the time. If you are reading this, you probably want to run the code. If this is the case, you need to download all of the files that were uploaded with this commit, along with "getMult.m" from an earlier commit. The file 'joel_filters.m" could be used in place of "adaptiveFilter.m" It was also uploaded in an earlier commit.